d3-finance-api/pyproject.toml
[project]
name = "d3-finance-api"
version = "0.1.0"
description = "Back-end projeto tcc proway"
authors = [
    {name = "Filipe Eduardo Wienhage",email = "filipe.eduw@hotmail.com"},
    {name = "Luis Henrique Spezia",email = "lhspezia777@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard] (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "sqlalchemy (>=2.0.40,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "pymysql (>=1.1.1,<2.0.0)",
    "pydantic (>=2.11.1,<3.0.0)",
    "email-validator (>=2.1.0,<3.0.0)",
    "click (>=8.1.8,<9.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "cryptography (>=44.0.2,<45.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "pyTelegramBotAPI (>=4.15.0,<5.0.0)",
    "aiohttp (>=3.9.0,<4.0.0)"
]

[tool.poetry]
packages = [{include = "d3_finance_api", from = "src"}]

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
flake8 = "^7.2.0"
uvicorn = {extras = ["standard"], version = "^0.34.1"}


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
